% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2.R
\name{hist2}
\alias{hist2}
\title{Titled histogram}
\usage{
hist2(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{further arguments and graphical parameters passed to graphics::hist function.}
}
\description{
A simple function to create a titled histogram
}
\examples{
hist2(1:10)

}
