% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyAppPackage.R
\docType{package}
\name{ShinyAppPackage}
\alias{ShinyAppPackage}
\alias{ShinyAppPackage-package}
\title{ShinyAppPackage: A Shiny application for generate a titled histogram}
\description{
The ShinyAppPackage package runs a Shiny application that allows to the user
to modify the number of the bins in a titled histogram
}
